Moving DatabaseInterface to its own file
